+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [0.2.1+1] - Apr 20, 2021
+* Minor Code Fixes.
+
+## [0.2.1] - Apr 13, 2021
+* Minor API Changes.
+
 ## [0.2.0] - Apr 10, 2021
 * Migration to null-safety.
