+++ example/assets/index.html
@@ -12,6 +12,9 @@
     <br>
     Please tap the text to see Javascript execution.
 </div>
+
+<img src="https://wallpapercave.com/wp/wp49731.jpg" alt="W3Schools.com">
+
 <script src="script.js"></script>
 </body>
 </html>
