+++ lib/webview_flutter_plus.dart
@@ -1,5 +1,5 @@
 library webview_flutter_plus;
 
 export 'package:webview_flutter/webview_flutter.dart'
-    hide WebView, WebViewController, WebViewCreatedCallback;
+    hide WebView, WebViewCreatedCallback;
 export 'package:webview_flutter_plus/src/webview_flutter_plus.dart';
