+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: webview_flutter_plus
 description: An extension of webview_flutter to load local HTML,CSS and Javascript from Assets or Strings and much more.
-version: 0.2.0
+version: 0.2.1+1
 homepage: https://github.com/shah-xad/webview_flutter_plus
 
 environment:
